Turns out namespaces are a bad idea
